WIP, full api specs green with translationsV2 flag on and off
