+++ app/api/csv/specs/csvLoaderFixtures.ts
@@ -2,6 +2,7 @@ import db from 'api/utils/testing_db';
 import { propertyTypes } from 'shared/propertyTypes';
 import { templateUtils } from 'api/templates';
 import { LanguageISO6391 } from 'shared/types/commonTypes';
+import { fixturesTranslationsV2ToTranslationsLegacy } from 'api/i18n/specs/fixturesTranslationsV2ToTranslationsLegacy';
 
 const template1Id = db.id();
 const multiSelectThesaurusId = db.id();
@@ -173,11 +174,11 @@ export default {
     },
   ],
 
-  translations_v2: [
+  translations: fixturesTranslationsV2ToTranslationsLegacy([
     ...commonTranslationsV2('en'),
     ...commonTranslationsV2('es'),
     ...commonTranslationsV2('fr'),
-  ],
+  ]),
 };
 
 export { template1Id, templateWithGeneratedTitle, thesauri1Id };
