+++ app/api/utils/testingTenants.ts
@@ -1,3 +1,4 @@
+import { config } from 'api/config';
 import { Tenant, tenants } from 'api/tenants/tenantContext';
 
 const originalCurrentFN = tenants.current.bind(tenants);
@@ -7,6 +8,7 @@ let mockedTenant: Partial<Tenant>;
 const testingTenants = {
   mockCurrentTenant(tenant: Partial<Tenant>) {
     mockedTenant = tenant;
+    mockedTenant.featureFlags = mockedTenant.featureFlags || config.defaultTenant.featureFlags;
     tenants.current = () => <Tenant>mockedTenant;
   },
